Moved session store logic into a separate module
